requirements.txt
fastapi
uvicorn
pillow
pydantic
python-multipart
